% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_review.R
\name{export_review}
\alias{export_review}
\title{Export the matching results for manual review}
\usage{
export_review(
  matches,
  file,
  top_n = 5,
  n_reviewers = 3,
  date_format = "yyyy-mm-dd"
)
}
\arguments{
\item{matches}{The matching results as produced by \code{match_rows}, with
the \code{output} argument set to "review"}

\item{file}{The file path for the review document}

\item{top_n}{The number of matches to be included per row in the manual review}

\item{n_reviewers}{The number of reviewer columns to include}

\item{date_format}{The date format to be written to excel}
}
\description{
Export the matching results for manual review
}
\author{
Finlay Campbell (\email{finlaycampbell93@gmail.com})
}
