% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_review.R
\name{import_review}
\alias{import_review}
\title{Import the manual review of matches}
\usage{
import_review(
  file,
  dictionary = list(accept = c("accept", "yes"), reject = c("reject", "no"), discuss =
    c("discuss", "maybe"))
)
}
\arguments{
\item{file}{The file path to the completed review document, with the
"reviewer_" columns filled out}

\item{dictionary}{Map the values of the "reviewer_" columns to the accepted
outcomes "accept", "reject" or "discuss". This can be a named list
indicating which results map to which of the three outcome, or a function
accepting a character vector and outputting a character vector of outcomes
of the same length.}
}
\description{
Import the manual review of matches
}
\author{
Finlay Campbell (\email{finlaycampbell93@gmail.com})
}
