% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_dictionary.R
\name{match_dictionary}
\alias{match_dictionary}
\title{Update values by reference to dictionary.}
\usage{
match_dictionary(value, dictionary)
}
\arguments{
\item{value}{A vector of values to be updated.}

\item{dictionary}{A dictionary as generated by \code{update_dictionary}.}
}
\description{
Update values by reference to dictionary.
}
\author{
Finlay Campbell (\email{finlaycampbell93@gmail.com})
}
