% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_from_review.R
\name{merge_from_review}
\alias{merge_from_review}
\title{Merge two databases on the basis of manual reviews.}
\usage{
merge_from_review(x, y, review, ...)
}
\arguments{
\item{x}{The first database to be merged}

\item{y}{The second database to be merged}

\item{review}{The list returned by \code{import_review} or a string
indicating the file path of the review document}

\item{...}{Additional arguments passed to \code{import_review}}
}
\description{
Merge two databases on the basis of manual reviews.
}
\author{
Finlay Campbell (\email{finlaycampbell93@gmail.com})
}
