% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_linelists.R
\docType{data}
\name{sample_linelists}
\alias{sample_linelists}
\title{Two sample linelists for case matching and merging}
\format{
An object of class \code{list} of length 2.
}
\usage{
sample_linelists
}
\description{
A list with two sample linelists (linelist_a and linelist_b).
}
\author{
Finlay Campbell \email{finlaycampbell93@gmail.com}
}
\keyword{datasets}
